requirements.txt

feedparser>=6.0.11
requests>=2.32.3
beautifulsoup4>=4.12.3
PyYAML>=6.0.2
python-dotenv>=1.0.1
SQLAlchemy>=2.0.32
pandas>=2.2.2
tenacity>=8.3.0
tzdata>=2024.1
